@@ -21,11 +21,11 @@ kapt {
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion '28.0.3'
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 28
         multiDexEnabled true
         versionCode 122
         versionName "2018.4.1"
@@ -235,14 +235,15 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
     implementation "org.jetbrains.anko:anko:$anko_version"
+    implementation 'androidx.core:core-ktx:1.0.1'
     //implementation 'androidx.core:core-ktx:1.0.0-alpha1'
 
     // Support libraries
     implementation 'com.android.support:multidex:1.0.3'
-    implementation "com.android.support:appcompat-v7:$support_version"
-    implementation "com.android.support:design:$support_version"
-    implementation "com.android.support:support-v4:$support_version"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation "com.google.android.material:material:1.1.0-alpha01"
+    implementation "androidx.legacy:legacy-support-v4:1.0.0"
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     // WorkManager
     implementation "android.arch.work:work-runtime-ktx:$work_version"
@@ -259,7 +260,13 @@ dependencies {
     //Room
     def room_version = "2.1.0-alpha02"
     implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
+
+    //Lifecycle
+    def lifecycle_version = "2.0.0"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
+    implementation "androidx.lifecycle:lifecycle-livedata:$lifecycle_version"
+    kapt "androidx.lifecycle:lifecycle-compiler:$lifecycle_version"
 
     // Firebase
     implementation "com.google.firebase:firebase-core:16.0.1"
