@@ -1,6 +1,6 @@
 package com.simprints.id
 
-import android.support.multidex.MultiDexApplication
+import androidx.multidex.MultiDexApplication
 import com.crashlytics.android.Crashlytics
 import com.crashlytics.android.core.CrashlyticsCore
 import com.simprints.id.data.db.DbManager
