@@ -2,7 +2,7 @@ package com.simprints.id.activities.checkLogin.openedByMainLauncher
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.simprints.id.Application
 import com.simprints.id.R
 import com.simprints.id.activities.dashboard.DashboardActivity
