@@ -1,6 +1,6 @@
 package com.simprints.id.activities.collectFingerprints
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import com.simprints.id.R
