@@ -1,7 +1,7 @@
 package com.simprints.id.activities.longConsent
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.text.method.ScrollingMovementMethod
 import android.view.View
 import com.simprints.id.Application
