@@ -2,7 +2,7 @@ package com.simprints.id.services.progress.notifications
 
 import android.app.Notification
 import android.app.NotificationManager
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
 import java.util.concurrent.atomic.AtomicBoolean
 
 
