@@ -1,7 +1,7 @@
 package com.simprints.id.services.progress.notifications
 
 import android.app.NotificationManager
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
 import com.simprints.id.services.progress.Progress
 import com.simprints.id.services.sync.SyncCategory
 import io.reactivex.observers.DisposableObserver
