@@ -3,7 +3,7 @@ package com.simprints.id.services.progress.notifications
 import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Intent
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
 import com.simprints.id.activities.dashboard.DashboardActivity
 import com.simprints.id.services.progress.Progress
 import com.simprints.id.services.sync.SyncCategory
