@@ -5,7 +5,7 @@ import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
 import android.os.Build
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
 import com.simprints.id.R
 import com.simprints.id.services.progress.Progress
 import com.simprints.id.services.progress.UploadProgress
