@@ -1,9 +1,6 @@
 package com.simprints.id.activities;
 
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 
 import com.simprints.id.Application;
 import com.simprints.id.R;
@@ -11,6 +8,10 @@ import com.simprints.id.data.prefs.PreferencesManager;
 import com.simprints.id.tools.LanguageHelper;
 
 import javax.inject.Inject;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 public class TutorialActivity extends AppCompatActivity {
 
