@@ -1,13 +1,12 @@
 package com.simprints.id.activities.collectFingerprints;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.animation.DecelerateInterpolator;
 import android.view.animation.Interpolator;
 import android.widget.Scroller;
-
 import java.lang.reflect.Field;
+import androidx.viewpager.widget.ViewPager;
 
 public class ViewPagerCustom extends ViewPager {
     private FixedSpeedScroller mScroller = null;
