@@ -1,9 +1,9 @@
 package com.simprints.id.activities.dashboard.views
 
-import android.support.v7.widget.RecyclerView
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.simprints.id.R
 import com.simprints.id.activities.dashboard.models.DashboardCard
 
