@@ -1,12 +1,12 @@
 package com.simprints.id.activities.dashboard.views
 
 import android.annotation.SuppressLint
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.CardView
 import android.view.View
 import android.widget.ImageView
 import android.widget.ProgressBar
 import android.widget.TextView
+import androidx.cardview.widget.CardView
+import androidx.core.content.ContextCompat
 import com.simprints.id.R
 import com.simprints.id.activities.dashboard.models.DashboardCard
 import com.simprints.id.activities.dashboard.models.DashboardSyncCard
