@@ -2,8 +2,8 @@ package com.simprints.id.activities.settings
 
 import android.os.Bundle
 import android.preference.PreferenceActivity
-import android.preference.PreferenceFragment
 import android.view.MenuItem
+import androidx.preference.PreferenceFragment
 import com.simprints.id.R
 import com.simprints.id.activities.settings.fragments.settingsPreference.SettingsPreferenceFragment
 import com.simprints.id.tools.extensions.isXLargeTablet
