@@ -3,7 +3,7 @@ package com.simprints.libcommon;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+
 import android.util.Log;
 
 import java.nio.BufferOverflowException;
@@ -13,6 +13,8 @@ import java.nio.ByteOrder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+
+import androidx.annotation.NonNull;
 
 public class Template {
     private byte[] bytes;
